+++ tests/html/MultipleTest.php
@@ -79,6 +79,42 @@ final class MultipleTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
+     * Parse multiple $quill_json strings and return via index, call via API,
+     * trim option set
+     */
+    public function testLoadingMultipleDeltasAndRenderingWithTrim()
+    {
+        $quill_json = [
+            'one' => $this->delta_one,
+            'two' => $this->delta_two
+        ];
+
+        $result_one = null;
+        $result_two = null;
+
+        try {
+            $quill = new RenderMultiple($quill_json, 'HTML');
+
+            $result_one = $quill->render('one', true);
+            $result_two = $quill->render('two', true);
+
+        } catch (\Exception $e) {
+            $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
+        }
+
+        $this->assertEquals(
+            $this->expected_one,
+            $result_one,
+            __METHOD__ . ' multiple text first index failure'
+        );
+        $this->assertEquals(
+            $this->expected_two,
+            $result_two,
+            __METHOD__ . ' multiple text second index failure'
+        );
+    }
+
+    /**
      * Parse multiple $quill_json strings, invalid index
      */
     public function testLoadingMultipleDeltasAndRenderingInvalidIndex()